move binary encoder to its own package
