+++ internal/testing/binaryencoding/import_test.go
@@ -1,9 +1,8 @@
-package binary
+package binaryencoding
 
 import (
 	"testing"
 
-	"github.com/tetratelabs/wazero/internal/testing/binaryencoding"
 	"github.com/tetratelabs/wazero/internal/testing/require"
 	"github.com/tetratelabs/wazero/internal/wasm"
 )
@@ -155,7 +154,7 @@ func TestEncodeImport(t *testing.T) {
 		tc := tt
 
 		t.Run(tc.name, func(t *testing.T) {
-			bytes := binaryencoding.EncodeImport(tc.input)
+			bytes := EncodeImport(tc.input)
 			require.Equal(t, tc.expected, bytes)
 		})
 	}
