+++ internal/wasm/binary/data.go
@@ -77,12 +77,3 @@ func decodeDataSegment(r *bytes.Reader, enabledFeatures api.CoreFeatures) (*wasm
 		Init:             b,
 	}, nil
 }
-
-func encodeDataSegment(d *wasm.DataSegment) (ret []byte) {
-	// Currently multiple memories are not supported.
-	ret = append(ret, leb128.EncodeInt32(0)...)
-	ret = append(ret, encodeConstantExpression(d.OffsetExpression)...)
-	ret = append(ret, leb128.EncodeUint32(uint32(len(d.Init)))...)
-	ret = append(ret, d.Init...)
-	return
-}
