+++ internal/wasm/binary/element.go
@@ -281,20 +281,3 @@ func decodeElementSegment(r *bytes.Reader, enabledFeatures api.CoreFeatures) (*w
 		return nil, fmt.Errorf("invalid element segment prefix: 0x%x", prefix)
 	}
 }
-
-// encodeCode returns the wasm.ElementSegment encoded in WebAssembly 1.0 (20191205) Binary Format.
-//
-// https://www.w3.org/TR/2019/REC-wasm-core-1-20191205/#element-section%E2%91%A0
-func encodeElement(e *wasm.ElementSegment) (ret []byte) {
-	if e.Mode == wasm.ElementModeActive {
-		ret = append(ret, leb128.EncodeInt32(int32(e.TableIndex))...)
-		ret = append(ret, encodeConstantExpression(e.OffsetExpr)...)
-		ret = append(ret, leb128.EncodeUint32(uint32(len(e.Init)))...)
-		for _, idx := range e.Init {
-			ret = append(ret, leb128.EncodeInt32(int32(*idx))...)
-		}
-	} else {
-		panic("TODO: support encoding for non-active elements in bulk-memory-operations proposal")
-	}
-	return
-}
