+++ internal/wasm/binary/table_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/tetratelabs/wazero/api"
+	"github.com/tetratelabs/wazero/internal/testing/binaryencoding"
 	"github.com/tetratelabs/wazero/internal/testing/require"
 	"github.com/tetratelabs/wazero/internal/wasm"
 )
@@ -54,7 +55,7 @@ func TestTableType(t *testing.T) {
 	for _, tt := range tests {
 		tc := tt
 
-		b := encodeTable(tc.input)
+		b := binaryencoding.EncodeTable(tc.input)
 		t.Run(fmt.Sprintf("encode - %s", tc.name), func(t *testing.T) {
 			require.Equal(t, tc.expected, b)
 		})
